bookstore/api-gateway/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
requests==2.31.0
PyJWT==2.8.0
redis==5.0.1
dj-database-url==2.1.0
django-cors-headers==4.3.1
gunicorn==21.2.0
